git commit -m "feat: 소셜 로그인 기능 추가로 인한 매물 관련 기능 수정 [S12P11D205-261]"

+++ backend/igeolu/src/main/java/com/ssafy/igeolu/facade/property/service/PropertyFacadeServiceImpl.java
@@ -31,6 +31,9 @@ import com.ssafy.igeolu.facade.property.dto.response.OptionListGetResponseDto;
 import com.ssafy.igeolu.facade.property.dto.response.PropertyGetResponseDto;
 import com.ssafy.igeolu.global.exception.CustomException;
 import com.ssafy.igeolu.global.exception.ErrorCode;
+import com.ssafy.igeolu.oauth.service.CustomOAuth2UserService;
+import com.ssafy.igeolu.oauth.service.SecurityService;
+import com.ssafy.igeolu.oauth.service.SecurityServiceImpl;
 import com.ssafy.igeolu.util.CoordinateConverter;
 
 import lombok.RequiredArgsConstructor;
@@ -48,9 +51,11 @@ public class PropertyFacadeServiceImpl implements PropertyFacadeService {
 	private final UserService userService;
 	private final FileService fileService;
 	private final PropertyOptionService propertyOptionService;
+	private final SecurityService securityService;
 
 	@Override
 	public void createProperty(PropertyPostRequestDto request, List<MultipartFile> images) {
+
 
 		// 좌표 변환
 		double[] latLon = coordinateConverter.convertToLatLon(
@@ -67,7 +72,8 @@ public class PropertyFacadeServiceImpl implements PropertyFacadeService {
 		// 옵션 ID 리스트로 Option 엔티티들 조회
 		List<Option> options = optionRepository.findByIdIn(request.getOptions());
 
-		User user = userService.getUserById(1);
+
+		User user = userService.getUserById(securityService.getCurrentUser().getUserId());
 
 		Property property = Property.builder()
 			.description(request.getDescription())
@@ -140,6 +146,21 @@ public class PropertyFacadeServiceImpl implements PropertyFacadeService {
 	@Override
 	public void updateProperty(Integer propertyId, PropertyUpdateRequestDto requestDto, List<MultipartFile> images) {
 		Property property = propertyService.getProperty(propertyId);
+
+		Integer currentUserId = securityService.getCurrentUser().getUserId();
+
+		// 매물 소유자와 현재 사용자가 다르면 예외 발생
+		if (!property.getUser().getId().equals(currentUserId)) {
+			throw new CustomException(ErrorCode.UNAUTHORIZED);
+		}
+
+
+		// dongcode 업데이트 추가
+		if (requestDto.getDongcode() != null) {
+			Dongcodes dongcode = dongcodesService.getDongcodes(requestDto.getDongcode());
+			property.setDongcode(dongcode);
+		}
+
 
 		// 좌표 변환
 		double[] latLon = coordinateConverter.convertToLatLon(
@@ -215,6 +236,13 @@ public class PropertyFacadeServiceImpl implements PropertyFacadeService {
 	@Override
 	public void deleteProperty(Integer propertyId) {
 		Property property = propertyService.getProperty(propertyId);
+
+		Integer currentUserId = securityService.getCurrentUser().getUserId();
+
+		// 매물 소유자와 현재 사용자가 다르면 예외 발생
+		if (!property.getUser().getId().equals(currentUserId)) {
+			throw new CustomException(ErrorCode.UNAUTHORIZED);
+		}
 
 		// 이미지 파일 삭제
 		property.getPropertyImages().forEach(i -> fileService.deleteFile(i.getFilePath()));
